change hardcoded value of course progress to actual value

+++ LMS/src/main/java/com/nt/LMS/dto/EnrolledUserDTO.java
@@ -9,7 +9,7 @@ public class EnrolledUserDTO {
     private Long userId;
     private String assignedByName;
     private String userName;
-    private Float progress;
+    private Double progress;
     private LocalDateTime enrollmentDate;
     private LocalDateTime deadline;
 }
